Fix parsing of submitForm command to find form correctly
The submitForm scripting command takes a single argument: the attribute
value and associated value used to identify the form in the DOM.
Currently, the logic for parsing this command attempts to use a
nonexistent (according to the syntax of the submitForm command) value to
identify the form instead of using the associated value of the
attribute. Fixed the parsing to use the appropriate value.