Set --emulated-form-factor=desktop for LH tests without device emulation
I believe the intention of setting --emulated-form-factor=none is to ask LH to run a "desktop" test, since the default value of this flag is "mobile". However, setting it to "none" seems to have some unintended side effects like preventing LH from adding the "Chrome-Lighthouse" user-agent modifier.

This would fix #296.