Ignore unicode errors in lighthouse output
Lighthouse is including unicode characters in some of the log lines, like an ellipsis. As a result, UnicodeDecodeErrors are thrown when lines with those unicode characters try to store themselves in self.task['lighthouse_log']. This change causes those characters to be ignored/thrown away so the lines can be properly represented in ascii. For Python3 support, any unicode() methods will need to be updated to use encode() and decode() methods.