feat: add logging by syslog
This change allows a user to opt into logging via syslog, in case they
want the agent to open a socket out to a service that consumes logs.

WARNING - This behavior only works on linux. It will _not_ work on MacOS,
since they changed the behavior of the syslog daemon, so we gate it for
linux